fix: apply Guardian compliance fixes (11 issues addressed)

+++ src/main/java/com/example/management/infrastructure/adapters/in/rest/dto/OrderItemResponse.java
@@ -2,23 +2,25 @@ package com.example.management.infrastructure.adapters.in.rest.dto;
 
 import com.example.management.application.dto.OrderItemDto;
 
+import java.math.BigDecimal;
+
 /**
  * DTO para un item de orden en las respuestas REST.
  */
 public record OrderItemResponse(
     String productId,
     String productName,
-    Double unitPrice,
+    BigDecimal unitPrice,
     Integer quantity,
-    Double total
+    BigDecimal total
 ) {
     public static OrderItemResponse fromApplicationDto(OrderItemDto itemDto) {
         return new OrderItemResponse(
             itemDto.productId(),
             itemDto.productName(),
-            itemDto.unitPrice().doubleValue(),
+            itemDto.unitPrice(),
             itemDto.quantity(),
-            itemDto.total().doubleValue()
+            itemDto.total()
         );
     }
 }
